+++ packages/sku/src/services/vite/index.ts
@@ -8,7 +8,7 @@ import { cleanTargetDirectory } from '@/utils/buildFileUtils.js';
 import { openBrowser } from '@/openBrowser/index.js';
 import { getAppHosts } from '@/utils/contextUtils/hosts.js';
 import chalk from 'chalk';
-import { startPrerenderWorkers } from '@/services/vite/helpers/prerenderRoutesWorker.js';
+import { startPrerenderWorkers } from '@/services/vite/helpers/prerender/prerenderConcurrently.js';
 import allocatePort from '@/utils/allocatePort.js';
 
 export const viteService = {
